prediction/requirements.txt
kafka-python~=2.0.2
pandas~=2.2.3
numpy~=2.0.0
scikit-learn~=1.5.1
joblib~=1.4.2
schedule~=1.2.2
pymongo~=4.10.1